created the UserService and put all the business logic of the user there
