+++ src/main/java/com/example/todo/controllers/UserController.java
@@ -1,0 +1,2 @@
+package com.example.todo.controllers;public class UserController {
+}
