--- src/main/java/com/example/todo/services/OTPService.java
@@ -1,11 +1,0 @@
-package com.example.todo.services;
-
-import org.springframework.stereotype.Service;
-
-@Service
-public class OTPService {
-
-    public void generateOTP(){}
-    public void sendOTP(){}
-    public void clearOTP(){}
-}
